Learning FourSquare API with Python
Introduction
In this lab, you will learn in details how to make calls to the Foursquare API for different purposes. You will learn how to construct a URL to send a request to the API to search for a specific type of venues, to explore a particular venue, to explore a Foursquare user, to explore a geographical location, and to get trending venues around a location. Also, you will learn how to use the visualization library, Folium, to visualize the results.

Table of Contents
Foursquare API Search Function
Explore a Given Venue
Explore a User
Foursquare API Explore Function
Get Trending Venues